On MSVC 2010 /Wall the CATCH_INTERNAL_LINEINFO macro was emitting the following warning:
  warning C4365: 'argument' : conversion from 'long' to 'size_t',
    signed/unsigned mismatch

To fix this I changed the __LINE__ usages to static_cast<std::size_t>(__LINE__).

It's kind of silly that MSVC is flagging this as a warning since it's a numeric
literal...

+++ include/internal/catch_common.h
@@ -118,12 +118,12 @@ namespace Catch {
     }
 }
 
-#define CATCH_INTERNAL_ERROR( msg ) throwLogicError( msg, __FILE__, __LINE__ );
+#define CATCH_INTERNAL_ERROR( msg ) throwLogicError( msg, __FILE__, static_cast<std::size_t>(__LINE__) );
 
 //#ifdef __FUNCTION__
-//#define CATCH_INTERNAL_LINEINFO ::Catch::SourceLineInfo( __FUNCTION__, __FILE__, __LINE__ )
+//#define CATCH_INTERNAL_LINEINFO ::Catch::SourceLineInfo( __FUNCTION__, __FILE__, static_cast<std::size_t>(__LINE__) )
 //#else
-#define CATCH_INTERNAL_LINEINFO ::Catch::SourceLineInfo( __FILE__, __LINE__ )
+#define CATCH_INTERNAL_LINEINFO ::Catch::SourceLineInfo( __FILE__, static_cast<std::size_t>(__LINE__) )
 //#endif
 
 #endif // TWOBLUECUBES_CATCH_COMMON_H_INCLUDED
